api/requirements.txt
aniso8601==8.1.0
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
JsonForm==0.0.2
jsonschema==3.2.0
JsonSir==0.0.2
MarkupSafe==1.1.1
nose==1.3.7
pyrsistent==0.17.3
python-dotenv==0.15.0
Python-EasyConfig==0.1.7
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.11
Werkzeug==1.0.1
